Make http rpc client optional
- add client feature and set as default
- Refactor http rpc code into client.rs
- maintain current exported symbols
- update ci and test for feature compatibility
