+++ .github/workflows/build.yml
@@ -22,6 +22,10 @@ jobs:
       - uses: actions-rs/cargo@v1
         with:
           command: check
+          args: --no-default-features
+      - uses: actions-rs/cargo@v1
+        with:
+          command: check
 
   clippy:
     name: Clippy
