+++ .github/workflows/test.yml
@@ -43,6 +43,12 @@ jobs:
           toolchain: stable
           override: true
 
+      - name: Cargo test (no default features)
+        uses: actions-rs/cargo@v1
+        with:
+          command: test
+          args: --no-default-features
+
       - name: Cargo test
         uses: actions-rs/cargo@v1
         with:
