+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "monero-rpc"
-version = "0.2.0"
+version = "0.2.1"
 authors = ["Artem Vorotnikov <artem@vorotnikov.me>", "Monero Rust Contributors"]
 edition = "2018"
 exclude = ["/test"]
@@ -10,19 +10,23 @@ repository = "https://github.com/monero-ecosystem/monero-rpc-rs"
 rust-version = "1.56.1"
 description = "RPC client for Monero daemon and wallet"
 
+[features]
+client = ["http", "jsonrpc-core", "reqwest", "tracing", "uuid"]
+default = ["client"]
+
 [dependencies]
 anyhow = "1"
 chrono = { version = "0.4", features = ["serde"] }
 fixed-hash = "0.7"
 hex = "0.4"
-http = "0.2"
-jsonrpc-core = "18"
+http = { version = "0.2", optional = true }
+jsonrpc-core = { version = "18", optional = true }
 monero = { version = "0.17", features = ["serde"] }
-reqwest = { version = "0.11", features = ["json"] }
+reqwest = { version = "0.11", features = ["json"], optional = true }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-tracing = "0.1"
-uuid = { version = "1.1", features = ["v4"] }
+tracing = { version = "0.1", optional = true }
+uuid = { version = "1.1", features = ["v4"], optional = true }
 
 [dev-dependencies]
 # Async
